+++ java-service-processor/src/test/resources/definition/expected/TestBatchReloadingLoader.java
@@ -89,6 +89,14 @@ public final class TestBatchReloadingLoader {
         return this;
       }
 
+      public <BACKEND> Builder backend(Function<Class<?>, BACKEND> factory,
+          Function<BACKEND, Iterable<?>> streamer) {
+        this.factory = (Function<Class<?>, Object>) factory;
+        this.streamer = (Function<Object, Iterable<?>>) streamer;
+        this.reloader = ignore -> {};
+        return this;
+      }
+
       public Mutable build() {
         Object providerBackend = factory.apply(TestBatchReloading.Mutable.class);
         Object batchBackend = factory.apply(TestBatchReloading.Batch.class);